requirements.txt
# Основные библиотеки для работы с данными
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.0.0

# Библиотеки для работы с языковыми моделями
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Библиотеки для мониторинга и логирования
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Библиотеки для работы с API и веб-сервисами
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv==1.0.0
pyyaml==6.0.1
hydra-core==1.3.2

# Библиотеки для работы с SIEM
elasticsearch>=8.8.0
python-whois>=0.8.0
dnspython>=2.4.0
pyOpenSSL>=23.2.0

# Утилиты
python-magic>=0.4.27
tldextract>=3.4.0
beautifulsoup4>=4.12.0
markdown>=3.4.0

# Добавленные библиотеки
torch>=2.0.0
scipy>=1.10.0 